cleaning up

+++ Robot_Arm/Servo_Joystick_PCA9685/Joint.cpp
@@ -1,5 +1,6 @@
 #include "Joint.hh"
 #include "HardwareSerial.h"
+
 #include <Adafruit_PWMServoDriver.h>
 #include <Arduino.h>
 #include <Wire.h>
@@ -9,8 +10,8 @@ extern HardwareSerial Serial;
 
 joint::joint() {}
 
-void joint::setJoint(byte pinNum, unsigned minVal, unsigned maxVal,
-                     unsigned pos, unsigned startPos) {
+void joint::setJoint(uint8_t pinNum, uint16_t minVal, uint16_t maxVal,
+                     uint16_t pos, uint16_t startPos) {
   this->pinNum = pinNum;
   this->maxVal = maxVal;
   this->minVal = minVal;
@@ -20,32 +21,18 @@ void joint::setJoint(byte pinNum, unsigned minVal, unsigned maxVal,
 
 void joint::move() { pwm.writeMicroseconds(this->pinNum, this->pos); }
 
-void joint::moveMs(unsigned time) {
+void joint::moveMs(uint16_t time) {
   this->pos = time;
   pwm.writeMicroseconds(this->pinNum, this->pos);
 }
 
-uint8_t joint::moveDegree(int angle) {
-  // Serial.print(angle);
-
-  // if (angle > 180)
-    // Serial.print("\t");
-    // Serial.println((pinNum + 7));
-  //   return ((pinNum + 1) * 10);
-  // if (angle < 0)
-    // Serial.print("\t");
-    // Serial.println(pinNum++);
-    // return pinNum++;
-
+uint8_t joint::moveDegree(uint8_t angle) {
   this->angle = angle;
   if (maxVal > 1800) {
     this->pos = map(angle, 0, 180, minVal, maxVal);
   } else {
     this->pos = map(angle, 0, 90, minVal, maxVal);
   }
-
-  // Serial.print("\t");
-  // Serial.println(pos);
 
   move();
