+++ Robot_Arm/Servo_Joystick_PCA9685/Joint.hh
@@ -1,5 +1,6 @@
 #ifndef JOINT_H
 #define JOINT_H
+
 #include <Adafruit_PWMServoDriver.h>
 #include <Arduino.h>
 #include <Wire.h>
@@ -17,19 +18,20 @@ private:
 public:
   joint();
 
-  void setJoint(byte pinNum, unsigned minVal, unsigned maxVal, unsigned pos,
-                unsigned startPos);
+  void setJoint(uint8_t pinNum, uint16_t minVal, uint16_t maxVal, uint16_t pos,
+                uint16_t startPos);
 
   void move();
 
-  void moveMs(unsigned time);
+  void moveMs(uint16_t time);
 
-  uint8_t moveDegree(int angle);
+  uint8_t moveDegree(uint8_t angle);
 
   void moveSteps(bool direction);
 
   unsigned position();
 };
+
 #endif // !JOINT_H
 
 // vim:filetype=cpp
